ragboard/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.27.0
aiofiles==23.2.1
boto3==1.34.14
pinecone-client==3.0.0
openai==1.8.0
anthropic==0.8.1
Pillow==10.2.0
PyPDF2==3.0.1
pytesseract==0.3.10
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
email-validator==2.1.0
asyncpg==0.29.0
chromadb==1.0.15
